requirements.txt
Flask==2.2.3
hazelcast-python-client==5.3
